video_like APIs finished

+++ src/main/java/com/braindevs/entity/VideoLikeEntity.java
@@ -19,7 +19,7 @@ public class VideoLikeEntity {
     private Integer id;
 
     @Column(name = "profile_id")
-    private Integer profileId;
+    private Long profileId;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "profile_id", insertable = false, updatable = false)
     private ProfileEntity profile;
